fix: deduplicate nursery data helper and use precise test assertions
