+++ tests/attacks/test_factory.py
@@ -50,6 +50,5 @@ def test_factory(monkeypatch, get_target):
     metrics = report[nr]["attack_experiment_logger"]["attack_instance_logger"][
         "instance_0"
     ]
-    # Synthetic nursery-like data can shift LiRA point estimates slightly.
-    assert metrics["TPR"] >= 0.9
-    assert 0.3 <= metrics["FPR"] <= 0.5
+    assert metrics["TPR"] == pytest.approx(0.95, abs=0.01)
+    assert metrics["FPR"] == pytest.approx(0.46, abs=0.01)
